+++ src/schema_first/openapi/schemas/v3_1_1/info_object_schema.py
@@ -4,15 +4,15 @@ from marshmallow import validate
 from ..base import BaseSchema
 from ..base import DocStringFields
 from ..constants import RE_VERSION
-from .contact_schema import ContactSchema
-from .license_schema import LicenseSchema
+from .contact_object_schema import ContactObjectSchema
+from .license_object_schema import LicenseObjectSchema
 
 
-class InfoSchema(DocStringFields, BaseSchema):
+class InfoObjectSchema(DocStringFields, BaseSchema):
     title = fields.String(required=True)
     version = fields.String(required=True, validate=validate.Regexp(RE_VERSION))
 
     termsOfService = fields.String()
 
-    contact = fields.Nested(ContactSchema)
-    license = fields.Nested(LicenseSchema)
+    contact = fields.Nested(ContactObjectSchema)
+    license = fields.Nested(LicenseObjectSchema)