+++ src/schema_first/openapi/schemas/v3_1_1/tag_object_schema.py
@@ -3,13 +3,13 @@ from marshmallow import validate
 
 from ..base import BaseSchema
 from ..base import DocStringFields
-from .external_docs_schema import ExternalDocsSchema
+from .external_docs_object_schema import ExternalDocsObjectSchema
 
 kinds = ['audience', 'badge', 'nav']
 
 
-class TagSchema(DocStringFields, BaseSchema):
+class TagObjectSchema(DocStringFields, BaseSchema):
     name = fields.String(required=True)
-    externalDocs = fields.Nested(ExternalDocsSchema)
+    externalDocs = fields.Nested(ExternalDocsObjectSchema)
     parent = fields.String()
     kind = fields.String(validate=validate.OneOf(choices=kinds))