+++ tests/converter/test_converter_boolean.py
@@ -9,7 +9,7 @@ from tests.utils import get_schema_from_request
 def test_converter__boolean_schema_fields(fx_spec_full, fx_spec_as_file):
     spec_file = fx_spec_as_file(fx_spec_full)
     spec = Specification(spec_file).load()
-    request_schema = get_schema_from_request(spec.reassembly_spec, '/endpoint', '200')
+    request_schema = get_schema_from_request(spec.reassembly_spec, '/full-endpoint', '200')
 
     assert isinstance(request_schema().fields['boolean'], fields.Boolean)
 
@@ -20,7 +20,7 @@ def test_converter__boolean_schema_fields(fx_spec_full, fx_spec_as_file):
 def test_converter__boolean_valid(fx_spec_full, fx_spec_as_file, schema):
     spec_file = fx_spec_as_file(fx_spec_full)
     spec = Specification(spec_file).load()
-    request_schema = get_schema_from_request(spec.reassembly_spec, '/endpoint', '200')
+    request_schema = get_schema_from_request(spec.reassembly_spec, '/full-endpoint', '200')
 
     schema['message'] = 'Valid string'
     request_schema().load(schema)
@@ -32,7 +32,7 @@ def test_converter__boolean_valid(fx_spec_full, fx_spec_as_file, schema):
 def test_converter__boolean_not_valid(fx_spec_full, fx_spec_as_file, schema):
     spec_file = fx_spec_as_file(fx_spec_full)
     spec = Specification(spec_file).load()
-    request_schema = get_schema_from_request(spec.reassembly_spec, '/endpoint', '200')
+    request_schema = get_schema_from_request(spec.reassembly_spec, '/full-endpoint', '200')
 
     schema['message'] = 'Valid string'
     with pytest.raises(ValidationError) as e:
@@ -45,7 +45,7 @@ def test_converter__boolean_not_valid(fx_spec_full, fx_spec_as_file, schema):
 def test_converter__string_nullable(fx_spec_full, fx_spec_as_file, schema):
     spec_file = fx_spec_as_file(fx_spec_full)
     spec = Specification(spec_file).load()
-    request_schema = get_schema_from_request(spec.reassembly_spec, '/endpoint', '200')
+    request_schema = get_schema_from_request(spec.reassembly_spec, '/full-endpoint', '200')
 
     schema['message'] = 'Valid string'
     request_schema().load(schema)