+++ tests/objects/info_object/conftest.py
@@ -8,13 +8,13 @@ def fx_info_object_required() -> dict:
 
 @pytest.fixture
 def fx_info_object_full(
-    fx_info_object_required, fx_contact_object_full, fx_license_object_full
+    fx_info_object_required, fx_contact_object__full, fx_license_object_full
 ) -> dict:
     return {
         **fx_info_object_required,
         'summary': 'A pet store manager.',
         'description': 'This is an example server for a pet store.',
         'termsOfService': 'https://example.com/terms/',
-        'contact': fx_contact_object_full,
+        'contact': fx_contact_object__full,
         'license': fx_license_object_full,
     }